+++ examples/resources/iosxe_interface_ethernet/resource.tf
@@ -45,4 +45,6 @@ resource "iosxe_interface_ethernet" "example" {
   ]
   arp_timeout             = 300
   spanning_tree_link_type = "point-to-point"
+  service_policy_input    = "POLICY1"
+  service_policy_output   = "POLICY1"
 }