+++ internal/provider/resource_iosxe_interface_ethernet_test.go
@@ -78,6 +78,8 @@ func TestAccIosxeInterfaceEthernet(t *testing.T) {
 	if os.Getenv("C9000V") != "" {
 		checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ip_dhcp_snooping_trust", "true"))
 	}
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "service_policy_input", "POLICY1"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "service_policy_output", "POLICY1"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -108,13 +110,20 @@ resource "iosxe_restconf" "PreReq0" {
 	}
 }
 
+resource "iosxe_restconf" "PreReq1" {
+	path = "Cisco-IOS-XE-native:native/policy/Cisco-IOS-XE-policy:policy-map=POLICY1"
+	attributes = {
+		"name" = "POLICY1"
+	}
+}
+
 `
 
 func testAccIosxeInterfaceEthernetConfig_minimum() string {
 	config := `resource "iosxe_interface_ethernet" "test" {` + "\n"
 	config += `	type = "GigabitEthernet"` + "\n"
 	config += `	name = "3"` + "\n"
-	config += `	depends_on = [iosxe_restconf.PreReq0, ]` + "\n"
+	config += `	depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, ]` + "\n"
 	config += `}` + "\n"
 	return config
 }
@@ -177,7 +186,9 @@ func testAccIosxeInterfaceEthernetConfig_all() string {
 	if os.Getenv("C9000V") != "" {
 		config += `	ip_dhcp_snooping_trust = true` + "\n"
 	}
-	config += `	depends_on = [iosxe_restconf.PreReq0, ]` + "\n"
+	config += `	service_policy_input = "POLICY1"` + "\n"
+	config += `	service_policy_output = "POLICY1"` + "\n"
+	config += `	depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, ]` + "\n"
 	config += `}` + "\n"
 	return config
 }